pyproject.toml
[project]
name = "multillmchatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.67.0",
    "cohere>=5.17.0",
    "google-generativeai>=0.8.5",
    "groq>=0.31.1",
    "openai>=1.107.1",
    "streamlit>=1.49.1",
]
